feat: Windows support for cuda-oxide
Port cuda-oxide from Linux-only to Windows (x86_64-pc-windows-msvc).

Changes:
- Fix MSVC bindgen enum type mismatches (i32 vs u32) in cuda-core
- Add PE/COFF 65535 export limit workaround via custom .def + build.rs
- Add COFF section support for PTX embedding in oxide-artifacts
- Add Windows .dll path detection in cargo-oxide backend discovery
- Use lld-link for codegen backend linking

Tested on Windows 11, RTX 4050 (SM_89), CUDA 13.1, Rust nightly-2026-04-03.

6 files changed, 57 insertions(+), 25 deletions(-)

+++ crates/cuda-core/src/stream.rs
@@ -100,7 +100,7 @@ impl CudaStream {
         let cu_stream = unsafe {
             cuda_bindings::cuStreamCreate(
                 cu_stream.as_mut_ptr(),
-                cuda_bindings::CUstream_flags_enum_CU_STREAM_NON_BLOCKING,
+                cuda_bindings::CUstream_flags_enum_CU_STREAM_NON_BLOCKING as u32,
             )
             .result()?;
             cu_stream.assume_init()
@@ -148,7 +148,7 @@ impl CudaStream {
             cuda_bindings::cuStreamWaitEvent(
                 self.cu_stream,
                 event.cu_event(),
-                cuda_bindings::CUevent_wait_flags_enum_CU_EVENT_WAIT_DEFAULT,
+                cuda_bindings::CUevent_wait_flags_enum_CU_EVENT_WAIT_DEFAULT as u32,
             )
             .result()
         }
